Add WFx tests and general cleanup

+++ aarch64/el1_ns/el1_nsec.c
@@ -11,6 +11,6 @@ void el1_init_el0()
 
     main = el1_load_el0((char *)EL0_NS_FLASH_BASE, (char *)EL0_NS_BASE_VA);
 
-    __set_exception_return(main);
+    __set_exception_return((uint64_t)main);
     __exception_return();
 }
